+++ src/qcodeedit/lib/document/qdocumentline.h
@@ -182,6 +182,7 @@ class QCE_EXPORT QDocumentLine
 		bool hasOverlay(int id);
 		QList<QFormatRange> getOverlays(int preferredFormat = -1) const;
 		QFormatRange getOverlayAt(int index, int preferredFormat = -1) const;
+        QFormatRange getOverlayAt(int index, QList<int> preferredFormats) const;
 		QFormatRange getFirstOverlay(int start = 0, int end = -1, int preferredFormat = -1) const;
 		QFormatRange getLastOverlay(int start = 0, int end = -1, int preferredFormat = -1) const;