+++ utilities/manual/source/CHANGELOG.md
@@ -2,6 +2,9 @@
 ## TeXstudio 4.8.6
 
 - provide command completion for expl3 commands
+- hide grammar errors in math/non-text environments (word repetition and language tool)
+- hide spelling errors in picture envs like tikzpicture
+- better handle \newtcbox [#3878](https://github.com/texstudio-org/texstudio/issues/3878)
 
 ## TeXstudio 4.8.5